Export api methods to module
First step for implementation of the complete hetzner dns api (https://dns.hetzner.com/api-docs)
